requirements.txt
aiogram==2.8
aiohttp==3.6.2
appdirs==1.4.4
arsenic==19.1
asgiref==3.2.7
async-timeout==3.0.1
attrs==19.3.0
Babel==2.8.0
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.6
emoji==0.5.4
emot==2.1
fake-useragent==0.1.11
idna==2.9
multidict==4.7.6
pyee==7.0.2
pyppeteer==0.2.2
pytz==2020.1
requests==2.23.0
selenium==3.141.0
six==1.15.0
sqlparse==0.3.1
structlog==20.1.0
tqdm==4.46.0
urllib3==1.25.9
useragent==0.1.1
websockets==8.1
yandex-translater==6.0
yandex.translate==0.3.5
yarl==1.4.2
